java 17 compile issues
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1902761 13f79535-47bb-0310-9956-ffa450edef68

+++ poi-scratchpad/src/main/java/org/apache/poi/hemf/record/emfplus/HemfPlusFont.java
@@ -102,7 +102,7 @@ public class HemfPlusFont {
 
             // A string of Length Unicode characters that contains the name of the font family.
             family = StringUtil.readUnicodeLE(leis, len);
-            size += len*LittleEndianConsts.SHORT_SIZE;
+            size += (long) len * LittleEndianConsts.SHORT_SIZE;
 
             return size;
         }
